backend/requirements.txt
Flask 
openai 
python-dotenv 
PyPDF2
gunicorn